+++ src/keys/twilio.ts
@@ -1,3 +1,3 @@
-const accountSid = 'AC7feb1563904b5ec34a73f25b1384b171';
-const authToken = '349c51f8ed6f3a8cffa6de85f527743a';
+const accountSid = '<INSERT HERE>';
+const authToken = '<INSERT HERE>';
 export {accountSid, authToken};
