+++ src/utils/scheduling.ts
@@ -2,7 +2,7 @@ import express from 'express';
 import { Message, IMessage } from '../models/message.model';
 import { Outcome, IOutcome } from '../models/outcome.model';
 import schedule from 'node-schedule';
-import {accountSid, authToken} from '../routes/keys';
+import { accountSid, authToken } from '../keys/twilio';
 
 const twilio = require('twilio')(accountSid, authToken);
